Restructure to add subpages
Will allow bringing in the pokedex
