+++ src/components/ColumnBody.tsx
